[refactor] 회원 탈퇴 기능 리팩토링

+++ src/main/java/com/ssafy/trip/member/service/MemberServiceImpl.java
@@ -75,14 +75,18 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public int delete(String id) {
+    public boolean delete(String id, String password) {
         try {
+            Member member = getMemberById(id);
+            if (!getDigest(id, password).equals(member.getPassword())) {
+                return false;
+            }
             memberMapper.deleteMember(id);
         } catch (Exception ex) {
             System.out.println(ex.getMessage());
-            return 0;
+            return false;
         }
-        return 1;
+        return true;
     }
 
     @Override
